refactor(orchestrator): extract events + IO from runner.go
Last sub-PR of the runner.go split. Pull the event emission + step/
artifact persistence + run-resume / git-summary / failed-finalize
helpers into a sibling so runner.go becomes a focused lifecycle
file (StartTask / ResumeTask / RetryTask / ContinueTask / CancelTask
+ Runner construction + the executeRun loop).

  runner.go:     1574 → 1233 LOC
  runner_io.go:  new, 334 LOC

What moves to runner_io.go

  Event emission
  - emitRunEvent: the single helper every other runner path goes
    through to write a TaskRunEvent + side-channel telemetry.

  Step + artifact upsert
  - upsertStep / upsertArtifact: bounded retry around the
    store-level Upsert, with the conflict-vs-genuine-failure split
    the runner cares about.

  Resume + retry plumbing
  - resumeCheckpointForRun: locates the last completed step + last
    event sequence for a prior run, so a retry-from-turn or
    resume-after-approval can start from there instead of replaying
    the whole conversation.
  - finalizeFailedRun: the canonical "stop the run, emit run.failed
    or run.cancelled, clear in-flight refs" path that every fail
    branch lands in.

  Git summary artifact
  - gitSummaryArtifact: synthesizes the run's git-diff summary
    artifact when the workspace looks like a git repo on a real
    branch with changes; called from the run-finalize path.

What stays in runner.go

  Runner construction (NewRunner + dependency wiring), lifecycle
  (StartTask, ResumeTask, RetryTask, ContinueTask, CancelTask and
  the *FromTurn / startTaskWithOptions variants), the executeRun
  loop that drives one run through its executor, the
  RuntimeStats reporter, and the small id/helper utilities
  (defaultResourceID, recordOrchestratorRunStarted, etc.).

The receiver type stays *Runner; no public API change; no test
changes. Imports per file pruned via goimports.

Verification

  go build ./...                                — clean
  go vet ./...                                  — clean
  gofmt -s -l .                                 — clean
  go test ./internal/orchestrator
    -run 'TestRunner|runner_*' -race -count=5   — 50 / 50 pass
  go test ./... -race -count=1                  — 1779 / 1779 pass
